replace interface where there is just one impl
